add hybrid gauge for new current method
