+++ source/module_hamilt_lcao/module_tddft/CMakeLists.txt
@@ -11,7 +11,6 @@ if(ENABLE_LCAO)
         td_velocity.cpp
         td_current.cpp
         snap_psibeta_half_tddft.cpp
-        td_folding_mixing.cpp
     )
 
     add_library(
